created a seperate file for players class
